Conectando con endpoint de registro

+++ RoomiesU_Backend/src/main/java/com/roomies/roomiesu/controller/request/CreateUsuarioDTO.java
@@ -26,12 +26,12 @@ public class CreateUsuarioDTO {
     @NotBlank(message = "La contraseña es obligatoria")
     private String password;
 
-    private Integer identificacion;
+    private Long identificacion;
 
     @Email(message = "Debe ser un email válido")
     private String email;
 
-    private Integer telefono;
+    private Long telefono;
     private Integer edad;
     private String descripcion;
     private Boolean isAvailable;
